fix tokenset tree breaking (#1598)
