--- src/app/components/TokenSetListOrTree/index.ts
@@ -1,2 +1,0 @@
-export * from './TokenSetListOrTree';
-export * from './TokenSetListOrTreeDragItem';
