+++ src/app/components/TokenSetSelector.tsx
@@ -49,20 +49,15 @@ export default function TokenSetSelector({ saveScrollPositionSet }: { saveScroll
   const [showRenameTokenSetFields, setShowRenameTokenSetFields] = React.useState(false);
   const [newTokenSetName, handleNewTokenSetNameChange] = React.useState('');
   const [tokenSetMarkedForChange, setTokenSetMarkedForChange] = React.useState('');
-  const [allTokenSets, setAllTokenSets] = React.useState(Object.keys(tokens));
-  const tokenKeys = Object.keys(tokens).join(',');
+  const allTokenSets = React.useMemo(() => Object.keys(tokens), [tokens]);
 
   React.useEffect(() => {
-    const scollPositionSet = allTokenSets.reduce<Record<string, number>>((acc, crr) => {
+    const scrollPositionSet = allTokenSets.reduce<Record<string, number>>((acc, crr) => {
       acc[crr] = 0;
       return acc;
     }, {});
-    dispatch.uiState.setScrollPositionSet(scollPositionSet);
+    dispatch.uiState.setScrollPositionSet(scrollPositionSet);
   }, [allTokenSets, dispatch]);
-
-  React.useEffect(() => {
-    setAllTokenSets(Object.keys(tokens));
-  }, [tokenKeys]);
 
   React.useEffect(() => {
     setShowNewTokenSetFields(false);
