+++ src/utils/tokenset/ensureFolderIsTogether.ts
@@ -8,7 +8,7 @@ export function ensureFolderIsTogether<T extends TreeItem>(value: T, order: Item
     const originalIndex = order.findIndex((item) => item.value === value);
     const nextIndex = nextOrder.findIndex((item) => item.value === value);
     const delta = nextIndex - originalIndex;
-    const itemsToMove = nextOrder.filter((item) => item.value.parent?.startsWith(value.path));
+    const itemsToMove = nextOrder.filter((item) => item.value.parent === value.path);
     itemsToMove.forEach((itemToMove) => {
       if (delta !== 0) {
         const itemToMoveIndex = nextOrder.findIndex((item) => item === itemToMove);
